git-svn-id: svn+ssh://jop91.astron.nl/auto/home/jive_cvs/SOFTC/svn/trunk/sfxc@328 5736e3ea-d426-0410-8aaf-93e336ac3372

+++ include/Multiple_data_readers_controller.h
@@ -48,7 +48,7 @@ public:
   bool initialised(unsigned int i);
   
   /** The number of input readers **/
-  unsigned int number_of_data_readers();
+  size_t number_of_data_readers();
 private:
 
   void add_data_reader(int i, boost::shared_ptr<Data_reader> reader);
